Statistics part2
Problem Statement
Task 1:
1. Blood glucose levels for obese patients have a mean of 100 with a standard deviation of 15. A
researcher thinks that a diet high in raw cornstarch will have a positive effect on blood glucose
levels. A sample of 36 patients who have tried the raw cornstarch diet have a mean glucose
level of 108. Test the hypothesis that the raw cornstarch had an effect or not.
2. In one state, 52% of the voters are Republicans, and 48% are Democrats. In a second state,
47% of the voters are Republicans, and 53% are Democrats. Suppose a simple random sample
of 100 voters are surveyed from each state.
What is the probability that the survey will show a greater percentage of Republican voters in
the second state than in the first state?
3. You take the SAT and score 1100. The mean score for the SAT is 1026 and the standard
deviation is 209. How well did you score on the test compared to the average test taker?
Task 2:
1. Is gender independent of education level? A random sample of 395 people were surveyed
and each person was asked to report the highest education level they obtained. The data that
resulted from the survey is summarized in the following table:
High School Bachelors Masters Ph.d. Total
Female 60 54 46 41 201
Male 40 44 53 57 194
Total 100 98 99 98 395
Question: Are gender and education level dependent at 5% level of significance? In other
words, given the data collected above, is there a relationship between the gender of an
individual and the level of education that they have obtained?
2. Using the following data, perform a oneway analysis of variance using α=.05. Write up the
results in APA format.
[Group1: 51, 45, 33, 45, 67]
[Group2: 23, 43, 23, 43, 45]
[Group3: 56, 76, 74, 87, 56]
3. Calculate F Test for given 10, 20, 30, 40, 50 and 5,10,15, 20, 25.
For 10, 20, 30, 40, 50: